Get first test of 3x3 sym eigendecomp passing
